+++ unittests/db_to_kv_tests.cpp
@@ -1,6 +1,6 @@
 #include <eosio/testing/tester.hpp>
 #include <eosio/chain/backing_store/db_key_value_format.hpp>
-#include <eosio/chain/backing_store/db_key_value_iter_cache.hpp>
+#include <eosio/chain/backing_store/db_key_value_iter_store.hpp>
 
 #include <boost/test/unit_test.hpp>
 #include <cstring>
